MoviewAdapter 분리
